[refactor] 초기 설정값 저장소 위치 변경
