+++ app/src/main/java/com/ggaebiz/ggaebiz/data/datastore/ConfigDataStore.kt
@@ -13,17 +13,12 @@ class ConfigDataStore(private val dataStore: DataStore<Preferences>) {
     private val isVibrationStatus = booleanPreferencesKey("is_vibration_status")
     private val vibrationValue = intPreferencesKey("vibration_value")
     private val volumeValue = intPreferencesKey("volume_value")
-    private val isHomeNudgeGuideViewed = booleanPreferencesKey("is_home_nudge_viewed")
-    private val isSettingNudgeGuideViewed = booleanPreferencesKey("is_setting_nudge_viewed")
-    private val isBatteryPopupViewed = booleanPreferencesKey("is_battery_popup_viewed")
+
 
     companion object {
         const val DEFAULT_IS_VIBRATION = true
         const val DEFAULT_VIBRATION = 3
         const val DEFAULT_VOLUME = 3
-        const val DEFAULT_IS_HOME_NUDGE_GUIDE_VIEWED = false
-        const val DEFAULT_IS_SETTING_NUDGE_GUIDE_VIEWED = false
-        const val DEFAULT_IS_BATTERY_POPUP_VIEWED = false
     }
 
     suspend fun setVibrationStatus(value: Boolean) {
@@ -61,40 +56,4 @@ class ConfigDataStore(private val dataStore: DataStore<Preferences>) {
             preferences[volumeValue] ?: DEFAULT_VOLUME
         }
     }
-
-    suspend fun setHomeNudgeGuideViewed(value: Boolean) {
-        dataStore.edit { preferences ->
-            preferences[isHomeNudgeGuideViewed] = value
-        }
-    }
-
-    fun getHomeNudgeGuideViewed(): Flow<Boolean> {
-        return dataStore.data.map { preferences ->
-            preferences[isHomeNudgeGuideViewed] ?: DEFAULT_IS_HOME_NUDGE_GUIDE_VIEWED
-        }
-    }
-
-    suspend fun setSettingNudgeGuideViewed(value: Boolean) {
-        dataStore.edit { preferences ->
-            preferences[isSettingNudgeGuideViewed] = value
-        }
-    }
-
-    fun getSettingNudgeGuideViewed(): Flow<Boolean> {
-        return dataStore.data.map { preferences ->
-            preferences[isSettingNudgeGuideViewed] ?: DEFAULT_IS_SETTING_NUDGE_GUIDE_VIEWED
-        }
-    }
-    suspend fun setBatteryPopupViewed(value: Boolean) {
-        dataStore.edit { preferences ->
-            preferences[isBatteryPopupViewed] = value
-        }
-    }
-
-    fun getBatteryPopupViewed(): Flow<Boolean> {
-        return dataStore.data.map { preferences ->
-            preferences[isBatteryPopupViewed] ?: DEFAULT_IS_BATTERY_POPUP_VIEWED
-        }
-    }
-
 }
