+++ app/src/main/java/com/ggaebiz/ggaebiz/data/repository/ConfigRepositoryImpl.kt
@@ -1,9 +1,6 @@
 package com.ggaebiz.ggaebiz.data.repository
 
 import com.ggaebiz.ggaebiz.data.datastore.ConfigDataStore
-import com.ggaebiz.ggaebiz.data.datastore.ConfigDataStore.Companion.DEFAULT_IS_BATTERY_POPUP_VIEWED
-import com.ggaebiz.ggaebiz.data.datastore.ConfigDataStore.Companion.DEFAULT_IS_HOME_NUDGE_GUIDE_VIEWED
-import com.ggaebiz.ggaebiz.data.datastore.ConfigDataStore.Companion.DEFAULT_IS_SETTING_NUDGE_GUIDE_VIEWED
 import com.ggaebiz.ggaebiz.data.datastore.ConfigDataStore.Companion.DEFAULT_IS_VIBRATION
 import com.ggaebiz.ggaebiz.data.datastore.ConfigDataStore.Companion.DEFAULT_VIBRATION
 import com.ggaebiz.ggaebiz.data.datastore.ConfigDataStore.Companion.DEFAULT_VOLUME
@@ -28,22 +25,5 @@ class ConfigRepositoryImpl(
     override suspend fun getVolumeValue(): Int =
         configDataStore.getVolumeValue().firstOrNull() ?: DEFAULT_VOLUME
 
-    override suspend fun setHomeNudgeGuideViewed(value: Boolean) =
-        configDataStore.setHomeNudgeGuideViewed(value)
-
-    override suspend fun getHomeNudgeGuideViewed(): Boolean =
-        configDataStore.getHomeNudgeGuideViewed().firstOrNull() ?: DEFAULT_IS_HOME_NUDGE_GUIDE_VIEWED
-
-    override suspend fun setSettingNudgeGuideViewed(value: Boolean) =
-        configDataStore.setSettingNudgeGuideViewed(value)
-
-    override suspend fun getSettingNudgeGuideViewed(): Boolean =
-        configDataStore.getSettingNudgeGuideViewed().firstOrNull() ?: DEFAULT_IS_SETTING_NUDGE_GUIDE_VIEWED
-
-    override suspend fun setBatteryPopupViewed(value: Boolean) =
-        configDataStore.setBatteryPopupViewed(value)
-
-    override suspend fun getBatteryPopupViewed(): Boolean =
-        configDataStore.getBatteryPopupViewed().firstOrNull() ?: DEFAULT_IS_BATTERY_POPUP_VIEWED
 
 }
