+++ app/src/main/java/com/ggaebiz/ggaebiz/domain/repository/ConfigRepository.kt
@@ -10,14 +10,4 @@ interface ConfigRepository {
 
     suspend fun setVolumeValue(value: Int)
     suspend fun getVolumeValue(): Int
-
-    suspend fun setHomeNudgeGuideViewed(value: Boolean)
-    suspend fun getHomeNudgeGuideViewed(): Boolean
-
-    suspend fun setSettingNudgeGuideViewed(value: Boolean)
-    suspend fun getSettingNudgeGuideViewed(): Boolean
-
-    suspend fun setBatteryPopupViewed(value: Boolean)
-    suspend fun getBatteryPopupViewed(): Boolean
-
 }
