+++ app/src/main/java/com/ggaebiz/ggaebiz/domain/repository/OnboardingRepository.kt
@@ -3,4 +3,13 @@ package com.ggaebiz.ggaebiz.domain.repository
 interface OnboardingRepository {
     suspend fun setIsOnboardingExposed(value : Boolean)
     suspend fun getIsOnboardingExposed() : Boolean
+
+    suspend fun setHomeNudgeGuideViewed(value: Boolean)
+    suspend fun getHomeNudgeGuideViewed(): Boolean
+
+    suspend fun setSettingNudgeGuideViewed(value: Boolean)
+    suspend fun getSettingNudgeGuideViewed(): Boolean
+
+    suspend fun setBatteryPopupViewed(value: Boolean)
+    suspend fun getBatteryPopupViewed(): Boolean
 }
